refactor(invitation): move Invitation.kt to model.activities
