+++ app/src/androidTest/java/com/github/se/studentconnect/ui/screen/home/HomeScreenQRScannerTest.kt
@@ -5,6 +5,7 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.navigation.compose.rememberNavController
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.se.studentconnect.model.activities.Invitation
 import com.github.se.studentconnect.model.event.Event
 import com.github.se.studentconnect.model.event.EventParticipant
 import com.github.se.studentconnect.model.event.EventRepository
@@ -104,9 +105,7 @@ class HomeScreenQRScannerTest {
             fromUserId: String
         ) {}
 
-        override suspend fun getInvitations(
-            userId: String
-        ): List<com.github.se.studentconnect.ui.screen.activities.Invitation> = emptyList()
+        override suspend fun getInvitations(userId: String): List<Invitation> = emptyList()
 
         override suspend fun acceptInvitation(eventId: String, userId: String) {}
