+++ app/src/androidTest/java/com/github/se/studentconnect/ui/screen/profile/FriendsListScreenAndroidTest.kt
@@ -9,11 +9,11 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.se.studentconnect.model.activities.Invitation
 import com.github.se.studentconnect.model.friends.FriendsRepository
 import com.github.se.studentconnect.model.user.User
 import com.github.se.studentconnect.model.user.UserRepository
 import com.github.se.studentconnect.ui.profile.FriendsListViewModel
-import com.github.se.studentconnect.ui.screen.activities.Invitation
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import org.junit.Before
