+++ app/src/main/java/com/github/se/studentconnect/model/activities/Invitation.kt
@@ -1,4 +1,4 @@
-package com.github.se.studentconnect.ui.screen.activities
+package com.github.se.studentconnect.model.activities
 
 import com.google.firebase.Timestamp
 import com.google.firebase.firestore.ServerTimestamp
