+++ app/src/main/java/com/github/se/studentconnect/model/event/EventRepositoryFirestore.kt
@@ -3,9 +3,9 @@
 package com.github.se.studentconnect.model.event
 
 import android.util.Log
+import com.github.se.studentconnect.model.activities.Invitation
+import com.github.se.studentconnect.model.activities.InvitationStatus
 import com.github.se.studentconnect.model.location.Location
-import com.github.se.studentconnect.ui.screen.activities.Invitation
-import com.github.se.studentconnect.ui.screen.activities.InvitationStatus
 import com.google.firebase.Firebase
 import com.google.firebase.auth.auth
 import com.google.firebase.firestore.DocumentReference
