+++ app/src/main/java/com/github/se/studentconnect/model/event/EventRepositoryLocal.kt
@@ -1,8 +1,8 @@
 // Portions of this code were generated with the help of Gemini
 package com.github.se.studentconnect.model.event
 
-import com.github.se.studentconnect.ui.screen.activities.Invitation
-import com.github.se.studentconnect.ui.screen.activities.InvitationStatus
+import com.github.se.studentconnect.model.activities.Invitation
+import com.github.se.studentconnect.model.activities.InvitationStatus
 import java.util.UUID
 
 /**
