+++ app/src/main/java/com/github/se/studentconnect/ui/screen/activities/ActivitiesViewModel.kt
@@ -2,6 +2,8 @@ package com.github.se.studentconnect.ui.screen.activities
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.github.se.studentconnect.model.activities.Invitation
+import com.github.se.studentconnect.model.activities.InvitationStatus
 import com.github.se.studentconnect.model.authentication.AuthenticationProvider
 import com.github.se.studentconnect.model.event.EventParticipant
 import com.github.se.studentconnect.model.event.EventRepository
