+++ app/src/test/java/com/github/se/studentconnect/model/user/UserRepositoryFirestoreTest.kt
@@ -1,6 +1,6 @@
 package com.github.se.studentconnect.model.user
 
-import com.github.se.studentconnect.ui.screen.activities.InvitationStatus
+import com.github.se.studentconnect.model.activities.InvitationStatus
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.Timestamp
