+++ app/src/test/java/com/github/se/studentconnect/ui/profile/FriendsListViewModelTest.kt
@@ -1,5 +1,6 @@
 package com.github.se.studentconnect.ui.profile
 
+import com.github.se.studentconnect.model.activities.Invitation
 import com.github.se.studentconnect.model.friends.FriendsRepository
 import com.github.se.studentconnect.model.user.User
 import com.github.se.studentconnect.model.user.UserRepository
@@ -459,8 +460,7 @@ class FriendsListViewModelTest {
     override suspend fun addInvitationToUser(eventId: String, userId: String, fromUserId: String) =
         Unit
 
-    override suspend fun getInvitations(userId: String) =
-        emptyList<com.github.se.studentconnect.ui.screen.activities.Invitation>()
+    override suspend fun getInvitations(userId: String) = emptyList<Invitation>()
 
     override suspend fun acceptInvitation(eventId: String, userId: String) = Unit
