+++ app/src/test/java/com/github/se/studentconnect/ui/profile/UserCardViewModelTest.kt
@@ -1,5 +1,6 @@
 package com.github.se.studentconnect.ui.profile
 
+import com.github.se.studentconnect.model.activities.Invitation
 import com.github.se.studentconnect.model.user.User
 import com.github.se.studentconnect.model.user.UserRepository
 import kotlinx.coroutines.Dispatchers
@@ -400,8 +401,7 @@ class UserCardViewModelTest {
     override suspend fun addInvitationToUser(eventId: String, userId: String, fromUserId: String) =
         Unit
 
-    override suspend fun getInvitations(userId: String) =
-        emptyList<com.github.se.studentconnect.ui.screen.activities.Invitation>()
+    override suspend fun getInvitations(userId: String) = emptyList<Invitation>()
 
     override suspend fun acceptInvitation(eventId: String, userId: String) = Unit
