+++ app/src/test/java/com/github/se/studentconnect/ui/profile/VisitorProfileViewModelTest.kt
@@ -1,12 +1,12 @@
 package com.github.se.studentconnect.ui.profile
 
 import com.github.se.studentconnect.R
+import com.github.se.studentconnect.model.activities.Invitation
 import com.github.se.studentconnect.model.authentication.AuthenticationProvider
 import com.github.se.studentconnect.model.friends.FriendsRepository
 import com.github.se.studentconnect.model.user.User
 import com.github.se.studentconnect.model.user.UserRepository
 import com.github.se.studentconnect.resources.TestResourceProvider
-import com.github.se.studentconnect.ui.screen.activities.Invitation
 import com.github.se.studentconnect.ui.screen.visitorprofile.FriendRequestStatus
 import com.github.se.studentconnect.ui.screen.visitorprofile.VisitorProfileViewModel
 import com.github.se.studentconnect.util.MainDispatcherRule
