+++ app/src/test/java/com/github/se/studentconnect/ui/screen/activities/ActivitiesViewModelTest.kt
@@ -1,5 +1,7 @@
 package com.github.se.studentconnect.ui.screen.activities
 
+import com.github.se.studentconnect.model.activities.Invitation
+import com.github.se.studentconnect.model.activities.InvitationStatus
 import com.github.se.studentconnect.model.authentication.AuthenticationProvider
 import com.github.se.studentconnect.model.event.Event
 import com.github.se.studentconnect.model.event.EventParticipant
